FE: [fix] 거래 중에는 계좌삭제 버튼이 활성화되지 않도록 수정

+++ BeRich_FE/src/pages/AfterLogin/UserInfoScreen.jsx
@@ -1,18 +1,20 @@
 import { Button, Text } from "@rneui/base";
-import { View } from "react-native";
+import { Alert, View } from "react-native";
 import { ButtonStyles } from "../../styles/Button.style";
 import { TextStyles } from "../../styles/Text.style";
 import { BoxStyles } from "../../styles/Box.style";
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons';
 import { DeleteAccountSelectBox, LogoutSelectBox, WithdrawSelectBox } from "../../components/SelectBox";
-import React, { useState } from "react";
+import React, { useContext, useState } from "react";
 import { getBalanceAPI } from "../../api/getBalanceAPI";
 import { useFocusEffect } from '@react-navigation/native';
 import { getUserInfoAPI } from "../../api/getUserInfoAPI";
+import { AppContext } from "../../contexts/AppContext";
 
 export default function UserInfoScreen({ navigation }) {
     const [userInfo, setUserInfo] = useState({ "accountNum": "정보 불러오는 중", "firstName": "", "lastName": "", "loginId": "" });
     const [balance, setBalance] = useState(' -');
+    const { state, setState } = useContext(AppContext);
 
     useFocusEffect(
         React.useCallback(() => {
@@ -46,6 +48,10 @@ export default function UserInfoScreen({ navigation }) {
                     <View style={[{ flexDirection: "row", justifyContent: "space-between" }, BoxStyles.AICenter]}>
                         <Text style={[TextStyles.Main]}>{balance}</Text>
                         {(balance !== ' -') ? <Button buttonStyle={[ButtonStyles.MainButton]} onPress={() => {
+                            if (state.statusData[0].status == 'ACTIVE') {
+                                Alert.alert('경고', '자동거래가 끝나지 않은 경우 계좌를 삭제할 수 없습니다.')
+                                return
+                            }
                             DeleteAccountSelectBox(setUserInfo, setBalance)
                         }}>계좌정보 삭제</Button> : null}
                     </View>
